backend/requirements.txt
# Core FastAPI and server dependencies
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.4

# Database dependencies
sqlalchemy==2.0.40
alembic==1.15.2
psycopg2-binary==2.9.10

# Authentication and security
python-jose[cryptography]==3.3.0
cryptography==44.0.3

# Environment and configuration
python-dotenv==1.1.0

# HTTP client for API calls
httpx==0.28.1
requests==2.32.3

# LLM Provider APIs
openai==1.82.0
google-generativeai==0.7.2
google-genai==1.15.0

# Payment processing
stripe==11.3.0
